CSS BUTTON AND NAV STYLE
Hi Kaki

You have worked on your CSS and I think the look and feel of your site is very nice.

There are a couple of things that need your attention though.

*

The first one is content... the about page for example is missing content and that may make testing later harder...

So for all pages try to complete the content.

Then, the p text (paragraphs) would best not be bold as that makes it a bit harder to read... so p text (set in the body) should be font weight 400 and now you have that set to 600 (CSS line 17)

H1, h2 etc can be bold, text should not be

*

HTML structure on the about page for skills and tools in not correct

So you would have the h3 headings e.g. 'relevant skills' and 'tools I use' but the content in it, if you add them in a list should be this structure

<ul>
<li>Creative thinking</li>
<li>xxx</li>
</ul>

That will create a list... now you have many different ul's instead

CSS file

line 2: error in the color code, you have ##191a26; and it should be #191a26; sonly 1 #
Line 23: as I mentioned in the call, remove all HTML ... this line should not be here
a element: you have a and a:hover in lines 53-63... please remove as you repeat them in line 70-90.
also related the the a element, make sure that the order is correct
a
a:visited
a:hover, a:focus
a:active

Always that order

Same for buttons and menus, everything that is essentially an a element

So

.button
.button:visited
.button:hover, .button:focus
.button:active

nav a
nav a:visited
nav a:hover, nav:focus
nav:active